# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TravelAI API
  version: 0.1.0
  description: |
    Canonical HTTP contract for TravelAI services.
    Paths are namespaced by service responsibility.
    Success envelope: `{ "success": true, "data": ..., "meta"?: ... }`
    Errors: RFC 7807 Problem Details where applicable.
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: api (catalog, booking)
  - url: http://localhost:3002
    description: identity
  - url: http://localhost:3003
    description: ai
...
paths:
...
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      operationId: authRegister
      servers:
        - url: http://localhost:3002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/auth/login:
    post:
      tags: [auth]
      summary: Login
      operationId: authLogin
      servers:
        - url: http://localhost:3002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "423":
          description: Account locked
  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token
      operationId: authRefresh
      servers:
        - url: http://localhost:3002
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token
      operationId: authLogout
      servers:
        - url: http://localhost:3002
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "204":
          description: No content
  /v1/auth/me:
    get:
      tags: [auth]
      summary: Current user profile
      operationId: authMe
      servers:
        - url: http://localhost:3002
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/change-password:
    post:
      tags: [auth]
      summary: Change password for authenticated user
      operationId: authChangePassword
      servers:
        - url: http://localhost:3002
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password changed; refresh tokens revoked
        "400":
          description: Validation error
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthSuccess:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            expiresIn:
              type: integer
            user:
              $ref: "#/components/schemas/User"
...